Fixed some bugs and temporarily bypassed some interface requirements.
This is part of an attempt to get code functional and clean up bugs/problems that were
not caught by the PHP linter. These changes are based on real-world tests using only
MySQL database connections. Note that the code is NOT yet functional, but is closer
than it was recently.

Also note that the most recent exception being thrown (presumably errantly) is the one
on line 257 of src/class/DatabaseModel.php, claiming that I am missing the "name" value
and therefore cannot build a valid DSN. The function call in question based on the
stack trace shows that I did, in fact, provide the database name variable.